adding slots functionality to dashboard and sidebar menu
