+++ apps/dashboard/src/app/dashboard/components/schools/schools.component.html
@@ -1,0 +1,167 @@
+<mat-drawer-container class="container2" hasBackdrop="true">
+  <mat-drawer #drawer class="sidenav2" mode="over" [(opened)]="opened" position="end" hasBackdrop="true">
+    <mat-icon class="close-drawer" (click)="drawer.toggle()">close</mat-icon>
+    <form [formGroup]="schoolForm" (ngSubmit)="addSchool()">
+      <div class="form_container">
+        <p class="title">Add new school</p>
+        <mat-form-field appearance="fill">
+          <mat-label>School name</mat-label>
+          <input matInput type="text" formControlName="schoolName" />
+        </mat-form-field>
+        <mat-form-field appearance="fill">
+          <mat-label>Geo location</mat-label>
+          <input
+            matInput
+            type="text"
+            formControlName="geoLocation"
+            [value]="addres"
+            (keydown.enter)="$event.preventDefault()"
+            placeholder="Ingrese dirección"
+            autocorrect="off"
+            autocomplete="off"
+            autocapitalize="off"
+            spellcheck="off"
+            #search
+          />
+        </mat-form-field>
+        <div class="forgot_password_container">
+          <p *ngIf="!mapVisible">Don’t find the location? <a (click)="toggleMap()">Click here</a></p>
+          <p *ngIf="mapVisible"><a (click)="toggleMap()">Hide map</a></p>
+        </div>
+        <!-- map -->
+        <agm-map *ngIf="mapVisible" [latitude]="latitude" [longitude]="longitude" [zoom]="zoom">
+          <agm-marker
+            [latitude]="latitude"
+            [longitude]="longitude"
+            [markerDraggable]="true"
+            (dragEnd)="markerDragEnd($event)"
+          ></agm-marker>
+        </agm-map>
+        <mat-progress-bar [mode]="'indeterminate'" *ngIf="isPublishing"> </mat-progress-bar>
+        <div>
+          <div class="right_container">
+            <h5>{{ addres }}</h5>
+          </div>
+        </div>
+        <!-- end map -->
+        <div class="codes">
+          <div class="secuential2">
+            <div class="teacher">
+              <mat-icon class="tCircle">circle</mat-icon>
+              <p class="roleTitle">Teacher code</p>
+              <p class="code">{{ tCode }}</p>
+            </div>
+            <div class="admin">
+              <mat-icon class="aCircle">circle</mat-icon>
+              <p class="roleTitle">Admin code</p>
+              <p class="code">{{ aCode }}</p>
+            </div>
+          </div>
+          <div class="secuential2">
+            <div class="student">
+              <mat-icon class="sCircle">circle</mat-icon>
+              <p class="roleTitle">Student code</p>
+              <p class="code">{{ sCode }}</p>
+            </div>
+            <div class="expert">
+              <mat-icon class="eCircle">circle</mat-icon>
+              <p class="roleTitle">Expert code</p>
+              <p class="code">{{ eCode }}</p>
+            </div>
+          </div>
+          <div class="secuential2">
+            <div class="parent">
+              <mat-icon class="pCircle">circle</mat-icon>
+              <p class="roleTitle">Parent code</p>
+              <p class="code">{{ pCode }}</p>
+            </div>
+          </div>
+        </div>
+        <button
+          type="submit"
+          class="publishSchool_button custom_button"
+          mat-raised-button
+          color="primary"
+          [disabled]="schoolForm.invalid"
+        >
+          Publish school
+          <mat-progress-bar *ngIf="isPublishing" mode="indeterminate"></mat-progress-bar>
+        </button>
+      </div>
+    </form>
+  </mat-drawer>
+
+  <div class="secuential">
+    <p class="segment-title">{{ title }}</p>
+    <p (click)="drawer.toggle()" class="add-school">+ Add school</p>
+  </div>
+
+  <!-- table begin -->
+  <div class="table_container" #epltable>
+    <div>
+      <mat-table mat-table [dataSource]="dataSource" matSort>
+        <!-- ID Column -->
+        <ng-container matColumnDef="name">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>SCHOOL NAME</th>
+          <td mat-cell *matCellDef="let row">
+            {{ row.name }}<br />
+            <p class="addres"><mat-icon class="marker">place</mat-icon>{{ row.addres }}</p>
+          </td>
+        </ng-container>
+
+        <!-- Progress Column -->
+        <ng-container matColumnDef="tcode">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>
+            <mat-icon class="tCircle">circle</mat-icon> Teacher code
+          </th>
+          <td mat-cell *matCellDef="let row">&nbsp;&nbsp;{{ row.tcode }}</td>
+        </ng-container>
+
+        <!-- Progress Column -->
+        <ng-container matColumnDef="scode">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>
+            <mat-icon class="sCircle">circle</mat-icon> Student code
+          </th>
+          <td mat-cell *matCellDef="let row">&nbsp;&nbsp;{{ row.scode }}</td>
+        </ng-container>
+
+        <!-- Name Column -->
+        <ng-container matColumnDef="pcode">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>
+            <mat-icon class="pCircle">circle</mat-icon> Parent code
+          </th>
+          <td mat-cell *matCellDef="let row">&nbsp;&nbsp;{{ row.pcode }}</td>
+        </ng-container>
+
+        <!-- Color Column -->
+        <ng-container matColumnDef="acode">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>
+            <mat-icon class="aCircle">circle</mat-icon> Admin code
+          </th>
+          <td mat-cell *matCellDef="let row">&nbsp;&nbsp;{{ row.acode }}</td>
+        </ng-container>
+
+        <ng-container matColumnDef="ecode">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header>
+            <mat-icon class="eCircle">circle</mat-icon> Expert code
+          </th>
+          <td mat-cell *matCellDef="let row">&nbsp;&nbsp;{{ row.ecode }}</td>
+        </ng-container>
+
+        <ng-container matColumnDef="see">
+          <th mat-header-cell *matHeaderCellDef mat-sort-header></th>
+          <td class="actionSee" mat-cell *matCellDef="let row">
+            <button mat-icon-button (click)="seeSchool(event, row.key)">
+              <mat-icon class="info" color="primary">chevron_right</mat-icon>
+            </button>
+          </td>
+        </ng-container>
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
+      </mat-table>
+      <mat-progress-bar *ngIf="isLoading" mode="indeterminate"></mat-progress-bar>
+      <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]"></mat-paginator>
+    </div>
+  </div>
+  <!-- table ends -->
+</mat-drawer-container>
