+++ apps/dashboard/src/app/dashboard/components/schools/schools.component.scss
@@ -2,9 +2,9 @@
 
 .container {
   position: absolute;
-  top: 0px;
-  bottom: 0px;
-  left: 0px;
+  top: 12px;
+  bottom: 12px;
+  left: 12px;
   right: 0;
 }
 
@@ -80,7 +80,6 @@
 .container2 {
   width: 100%;
   height: 100%;
-  border: 1px solid rgba(0, 0, 0, 0.5);
 }
 
 .sidenav-content2 {
