+++ apps/dashboard/src/app/dashboard/dashboard.component.html
@@ -1,1 +1,5 @@
-<new-pangea-sidebar-menu></new-pangea-sidebar-menu>
+<new-pangea-sidebar-menu>
+  <ng-container marcoBody>
+    <new-pangea-schools></new-pangea-schools>
+  </ng-container>
+</new-pangea-sidebar-menu>
